Refactor extraction functions and update TypeScript definitions
- Made the `options` parameter optional in several extraction functions to enhance flexibility.
- Removed the exclusion of JavaScript files from TypeScript compilation in `tsconfig.json`.
- Deleted the `index.d.ts` file as it was no longer needed.
- Added comprehensive tests for buffer extraction functionality, ensuring robust coverage for various file types.
- Cleaned up the codebase by removing outdated test files and improving type definitions for better clarity.
