+++ lib/types.ts
@@ -113,10 +113,3 @@ export interface Options {
    */
   includeAltText?: boolean | undefined;
 }
-
-export interface URLOptions extends Options {
-  /**
-   * Used with fromUrl, if set, rather than using the content-type from the URL request, will use the provided typeOverride.
-   */
-  typeOverride?: string | undefined;
-}
