--- test/extract.test.js
@@ -1,928 +1,0 @@
-import path from 'node:path';
-import { describe, it, expect } from 'vitest';
-import { fromBufferWithMime } from '../lib/index.js';
-
-async function fromFileWithPath(filePath, options) {
-  const mimeType = options?.typeOverride || mime.getType(filePath);
-  const text = await fromBufferWithMime(mimeType, filePath, options);
-  return text;
-}
-
-describe('textract', () => {
-  describe('for .csv files ', () => {
-    // is some oddness testing html files, not sure what the deal is
-
-    it('from csv files', (done) => {
-      const docPath = path.join(__dirname, 'files', 'csv.csv');
-      fromFileWithPath(docPath, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text.length).to.eql(18);
-        expect(text).to.eql('Foo,Bar Foo2,Bar2 ');
-        done();
-      });
-    });
-
-    it('it will extract text from csv files and insert newlines in the right places', (done) => {
-      const docPath = path.join(__dirname, 'files', 'csv.csv');
-      fromFileWithPath(docPath, { preserveLineBreaks: true }, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text.length).toEqual(18);
-        expect(text).toEqual('Foo,Bar\nFoo2,Bar2\n');
-        done();
-      });
-    });
-  });
-
-  describe('for .html files', () => {
-    // is some oddness testing html files, not sure what the deal is
-
-    it('will extract text from html files and insert newlines in the right places', (done) => {
-      const docPath = path.join(__dirname, 'files', 'test.html');
-      fromFileWithPath(docPath, { preserveLineBreaks: true }, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text.length).toEqual(80);
-        expect(text.substring(0, 80)).toEqual(
-          '\nThis is a\nlong string\nof text\nthat should get extracted\nwith new lines inserted',
-        );
-        done();
-      });
-    });
-
-    it('will extract text from html files', (done) => {
-      const docPath = path.join(__dirname, 'files', 'Google.html');
-      fromFileWithPath(docPath, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text.length).toEqual(869);
-        expect(text.substring(565, 620)).to.eql(
-          'you say next. Learn more No thanks Enable "Ok Google" I',
-        );
-        done();
-      });
-    });
-
-    it('will extract text from html files and preserve alt text when asked', (done) => {
-      const docPath = path.join(__dirname, 'files', 'test-alt.html');
-      fromFileWithPath(docPath, { includeAltText: true }, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text.length).toEqual(46);
-        expect(text).toEqual(' This is a paragraph that has an image inside ');
-        done();
-      });
-    });
-  });
-
-  describe('for .rss files', () => {
-    it('will extract text from rss files', (done) => {
-      const docPath = path.join(__dirname, 'files', 'rss.rss');
-      fromFileWithPath(docPath, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text.length).toEqual(5399);
-        expect(text.substring(0, 100)).toEqual(
-          ' FeedForAll Sample Feed RSS is a fascinating technology. The uses for RSS are expanding daily. Take ',
-        );
-        done();
-      });
-    });
-
-    it('will extract text from rss files and preserve line breaks', (done) => {
-      const docPath = path.join(__dirname, 'files', 'rss.rss');
-      fromFileWithPath(docPath, { preserveLineBreaks: true }, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text.length).toEqual(5534);
-        expect(text.substring(0, 100)).toEqual(
-          '\n FeedForAll Sample Feed\n RSS is a fascinating technology. The uses for RSS are expanding daily. Tak',
-        );
-        done();
-      });
-    });
-  });
-
-  describe('for .epub files', { timeout: 5000 }, () => {
-    it('will extract text from epub files', (done) => {
-      const docPath = path.join(
-        __dirname,
-        'files',
-        'Metamorphosis-jackson.epub',
-      );
-
-      fromFileWithPath(docPath, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text.length).toEqual(119329);
-        expect(text.substring(3000, 3500)).toEqual(
-          'dboard so that he could lift his head better; found where the itch was, and saw that it was covered with lots of little white spots which he didn\'t know what to make of; and when he tried to feel the place with one of his legs he drew it quickly back because as soon as he touched it he was overcome by a cold shudder. He slid back into his former position. "Getting up early all the time", he thought, "it makes you stupid. You\'ve got to get enough sleep. Other travelling salesmen live a life of lu',
-        );
-        done();
-      });
-    });
-
-    it('will extract text from epub files and preserve line breaks', (done) => {
-      const docPath = path.join(
-        __dirname,
-        'files',
-        'Metamorphosis-jackson.epub',
-      );
-
-      fromFileWithPath(docPath, { preserveLineBreaks: true }, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text.length).toEqual(119342);
-        expect(text.substring(3000, 3500)).toEqual(
-          'rds the headboard so that he could lift his head better; found where the itch was, and saw that it was covered with lots of little white spots which he didn\'t know what to make of; and when he tried to feel the place with one of his legs he drew it quickly back because as soon as he touched it he was overcome by a cold shudder.\nHe slid back into his former position. "Getting up early all the time", he thought, "it makes you stupid. You\'ve got to get enough sleep. Other travelling salesmen live a',
-        );
-        done();
-      });
-    });
-  });
-
-  describe('for .atom files', () => {
-    it('will extract text from atom files', (done) => {
-      const docPath = path.join(__dirname, 'files', 'atom.atom');
-      fromFileWithPath(docPath, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text.length).toEqual(26731);
-        expect(text.substring(0, 100)).toEqual(
-          ' @{}[]tag:theregister.co.uk,2005:feed/theregister.co.uk/data_centre/storage/ The Register - Data Cen',
-        );
-        done();
-      });
-    });
-
-    it('will extract text from atom files and preserve line breaks', (done) => {
-      const docPath = path.join(__dirname, 'files', 'atom.atom');
-      fromFileWithPath(docPath, { preserveLineBreaks: true }, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text.length).toEqual(27441);
-        expect(text.substring(0, 100)).toEqual(
-          '\n @{}[]tag:theregister.co.uk,2005:feed/theregister.co.uk/data_centre/storage/\n The Register - Data C',
-        );
-        done();
-      });
-    });
-  });
-
-  describe('for .rtf files', () => {
-    it('will extract text from rtf files', (done) => {
-      const docPath = path.join(__dirname, 'files', 'sample.rtf');
-      fromFileWithPath(docPath, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text.substring(144, 220)).toEqual(
-          "bit of hidden text. So we're going to end this paragraph here and go on to a",
-        );
-        done();
-      });
-    });
-
-    it('will extract when there are spaces in the name', (done) => {
-      const docPath = path.join(__dirname, 'files', 'sample rtf.rtf');
-      fromFileWithPath(docPath, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text.substring(144, 220)).toEqual(
-          "bit of hidden text. So we're going to end this paragraph here and go on to a",
-        );
-        done();
-      });
-    });
-
-    it('will extract text from actual rtf files with lines left in', (done) => {
-      const docPath = path.join(__dirname, 'files', 'sample.rtf');
-      fromFileWithPath(docPath, { preserveLineBreaks: true }, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text.substring(144, 230)).toEqual(
-          "bit of hidden text. So we're going to end this paragraph here and go on to a nice litt",
-        );
-        done();
-      });
-    });
-  });
-
-  describe('for .doc files', () => {
-    it('will extract text from actual doc files', (done) => {
-      const docPath = path.join(__dirname, 'files', 'doc.doc');
-      fromFileWithPath(docPath, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text.substring(0, 100)).toEqual(
-          ' Word Specification Sample Working Draft 04, 16 August 2002 Document identifier: wd-spectools-word-s',
-        );
-        done();
-      });
-    });
-
-    it('will extract text from actual doc files with spaces in the name', (done) => {
-      const docPath = path.join(__dirname, 'files', 'doc space.doc');
-      fromFileWithPath(docPath, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text.substring(0, 100)).toEqual(
-          ' Word Specification Sample Working Draft 04, 16 August 2002 Document identifier: wd-spectools-word-s',
-        );
-        done();
-      });
-    });
-
-    it('will not extract text from text files masquerading as doc files', (done) => {
-      const docPath = path.join(__dirname, 'files', 'notadoc.doc');
-      fromFileWithPath(docPath, (error, text) => {
-        expect(text).toBeNull();
-        expect(
-          error.toString().indexOf('does not appear to really be a .doc file'),
-        ).toEqual(36);
-        done();
-      });
-    });
-
-    it('will extract text from large .doc', (done) => {
-      const docPath = path.join(__dirname, 'files', 'sample.doc');
-      fromFileWithPath(docPath, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text.length).toEqual(32658);
-        done();
-      });
-    });
-
-    it('will extract text preserving line breaks without word wrap', (done) => {
-      const docPath = path.join(
-        __dirname,
-        'files',
-        'multiple-long-paragraphs.doc',
-      );
-      fromFileWithPath(docPath, { preserveLineBreaks: true }, (error, text) => {
-        expect(error).toBeNull();
-        expect(text.match(/\r\n|\n/g).length).toEqual(21);
-        done();
-      });
-    });
-  });
-
-  describe('for .xls files', () => {
-    it('will extract text', (done) => {
-      const docPath = path.join(__dirname, 'files', 'test.xls');
-      fromFileWithPath(docPath, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text.substring(0, 20)).toEqual('This,is,a,spreadshee');
-        done();
-      });
-    });
-
-    it('will extract text from multi-line files', (done) => {
-      const docPath = path.join(__dirname, 'files', 'test-multiline.xls');
-      fromFileWithPath(docPath, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text.substring(0, 40)).toEqual(
-          'This,is,a,spreadsheet,yay! And ,this,is,',
-        );
-        done();
-      });
-    });
-
-    it('will extract text from multi-line files and keep line breaks', (done) => {
-      const docPath = path.join(__dirname, 'files', 'test-multiline.xls');
-      fromFileWithPath(docPath, { preserveLineBreaks: true }, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text.substring(0, 40)).toEqual(
-          'This,is,a,spreadsheet,yay!\nAnd ,this,is,',
-        );
-        done();
-      });
-    });
-  });
-
-  describe('for .xlsx files', () => {
-    it('will extract text and numbers from XLSX files', (done) => {
-      const filePath = path.join(__dirname, 'files', 'pi.xlsx');
-      fromFileWithPath(filePath, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text).toEqual('This is the value of PI:,3.141592 ');
-        done();
-      });
-    });
-
-    it('will extract text from XLSX files with multiple sheets', (done) => {
-      const filePath = path.join(__dirname, 'files', 'xlsx.xlsx');
-      fromFileWithPath(filePath, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text.substring(49, 96)).toEqual(
-          'Color,Pattern,Sex,GeneralSizePotential,GeneralA',
-        );
-        done();
-      });
-    });
-
-    it('will error when input file is not an actual xlsx file', (done) => {
-      const filePath = path.join(__dirname, 'files', 'notaxlsx.xlsx');
-      fromFileWithPath(filePath, (error) => {
-        expect(error).not.toBeNull();
-        expect(error.message).toBeInstanceOf(String);
-        expect(error.message.substring(0, 43)).toEqual(
-          'Could not extract notaxlsx.xlsx, Error: PRN',
-        );
-        done();
-      });
-    });
-  });
-
-  describe('for .pdf files', () => {
-    it('will extract text from actual pdf files', (done) => {
-      const filePath = path.join(__dirname, 'files', 'pdf.pdf');
-      fromFileWithPath(filePath, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text).toEqual('This is a test. Please ignore.');
-        done();
-      });
-    });
-
-    it('will extract pdf text and preserve multiple lines', (done) => {
-      const filePath = path.join(__dirname, 'files', 'testpdf-multiline.pdf');
-      fromFileWithPath(
-        filePath,
-        { preserveLineBreaks: true },
-        (error, text) => {
-          expect(error).toBeNull();
-          expect(text).toBeInstanceOf(String);
-          expect(text).toEqual(
-            'This is a test,\nA multi-line test,\nLets hope it works',
-          );
-          done();
-        },
-      );
-    });
-
-    it("will error out when pdf file isn't actually a pdf", (done) => {
-      const filePath = path.join(__dirname, 'files', 'notapdf.pdf');
-      fromFileWithPath(filePath, (error, text) => {
-        expect(text).toBeNull();
-        expect(error).not.toBeNull();
-        expect(error.message).toBeInstanceOf(String);
-        expect(error.message.substring(0, 34)).toEqual(
-          'Error extracting PDF text for file',
-        );
-        done();
-      });
-    });
-
-    it('will properly handle multiple columns', (done) => {
-      const filePath = path.join(__dirname, 'files', 'two_columns.pdf');
-      fromFileWithPath(
-        filePath,
-        { preserveLineBreaks: true },
-        (error, text) => {
-          expect(error).toBeNull();
-          expect(text).toBeInstanceOf(String);
-          expect(
-            text.indexOf(
-              'Abstract— This work deals with a multi-cell topology based\non current-source converters based power cells.',
-            ) > 500,
-          ).toBe(true);
-          done();
-        },
-      );
-    });
-
-    it('can handle files with spaces in the name', (done) => {
-      const filePath = path.join(__dirname, 'files', 'two columns.pdf');
-      fromFileWithPath(
-        filePath,
-        { preserveLineBreaks: true },
-        (error, text) => {
-          expect(error).toBeNull();
-          expect(text).toBeInstanceOf(String);
-          expect(
-            text.indexOf(
-              'Abstract— This work deals with a multi-cell topology based\non current-source converters based power cells.',
-            ) > 500,
-          ).toBe(true);
-          done();
-        },
-      );
-    });
-
-    it('can handle manage PDFs with passwords', (done) => {
-      const filePath = path.join(
-        __dirname,
-        'files',
-        'pdf-example-password.original.pdf',
-      );
-      fromFileWithPath(
-        filePath,
-        { pdftotextOptions: { userPassword: 'test' } },
-        (error, text) => {
-          expect(error).toBeNull();
-          expect(text).toBeInstanceOf(String);
-          expect(text.substring(0, 200)).toEqual(
-            'Backup4all –backup solution for network environments Starting from version 2 it is easier to install Backup4all in a network environment. Network administrators can install Backup4all on a single comp',
-          );
-          done();
-        },
-      );
-    });
-
-    it('can handle manage PDFS with full-width Japanese characters', (done) => {
-      const filePath = path.join(__dirname, 'files', 'full-width-j.pdf');
-      fromFileWithPath(filePath, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text.replace(/ /g, '').substring(2685, 2900)).toEqual(
-          '＄％＆＇（）＊＋，－．／０１２３４５６７８９：；＜＝＞？＠ＡＢＣＤＥＦＧＨＩＪＫＬＭＮＯＰＱＲＳＴＵＶＷＸＹＺ［＼］＾＿｀ａｂｃｄｅｆｇｈｉｊｋｌｍｎｏｐｑｒｓｔｕｖｗｘｙｚ｛｜｝～｟｠｡｢｣､･ｦｧｨｩｪｫｬｭｮｯｰｱｲｳｴｵｶｷｸｹｺｻｼｽｾｿﾀﾁﾂﾃﾄﾅﾆﾇﾈﾉﾊﾋﾌﾍﾎﾏﾐﾑﾒﾓﾔﾕﾖﾗﾘﾙﾚﾛﾜﾝﾞﾟﾡﾢﾣﾤﾥﾦﾧﾨﾩﾪﾫﾬﾭﾮﾯﾰﾱﾲﾳﾴﾵﾶﾷﾸﾹﾺﾻﾼﾽﾾￂￃￄￅￆￇￊￋￌￍￎￏￒￓￔￕￖￗￚￛￜ￠￡￢￣￤￥￦F',
-        );
-        done();
-      });
-    });
-
-    // it( 'can handle arabic', function( done ) {
-    //   var filePath = path.join( __dirname, 'files', 'arabic.pdf' );
-    //   fromFileWithPath( filePath, function( error, text ) {
-    //     expect( error ).to.be.null;
-    //     expect( text ).to.be.a( 'string' );
-    //     expect( text.substring( 0, 200 ) ).to.eql( '' );
-    //     done();
-    //   });
-    // });
-  });
-
-  describe('for .docx files', () => {
-    it('will extract text from actual docx files', (done) => {
-      const filePath = path.join(__dirname, 'files', 'docx.docx');
-      fromFileWithPath(filePath, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text.substring(0, 20)).toEqual('This is a test Just ');
-        done();
-      });
-    });
-
-    it('will extract text from actual docx files and preserve line breaks', (done) => {
-      const filePath = path.join(__dirname, 'files', 'docx.docx');
-      fromFileWithPath(
-        filePath,
-        { preserveLineBreaks: true },
-        (error, text) => {
-          expect(error).toBeNull();
-          expect(text).toBeInstanceOf(String);
-          expect(text.substring(20, 40)).toEqual('so you know:\nLorem i');
-          done();
-        },
-      );
-    });
-
-    it('will extract text from actual docx files and preserve line breaks [line-breaks.docx]', (done) => {
-      const filePath = path.join(__dirname, 'files', 'line-breaks.docx');
-      fromFileWithPath(
-        filePath,
-        { preserveLineBreaks: true },
-        (error, text) => {
-          expect(error).toBeNull();
-          expect(text).toBeInstanceOf(String);
-          expect(text).toEqual(
-            'Paragraph follows\n\nLine break follows\n\nend\n\n',
-          );
-          done();
-        },
-      );
-    });
-
-    it("will error out when docx file isn't actually a docx", (done) => {
-      const filePath = path.join(__dirname, 'files', 'notadocx.docx');
-      fromFileWithPath(filePath, (error, text) => {
-        expect(text).toBeNull();
-        expect(error).not.toBeNull();
-        expect(error.message).toBeInstanceOf(String);
-        expect(error.message.substring(0, 34)).toEqual(
-          'File not correctly recognized as z',
-        );
-        done();
-      });
-    });
-
-    it('will not extract smashed together text', (done) => {
-      const filePath = path.join(__dirname, 'files', 'testresume.docx');
-      fromFileWithPath(filePath, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text.substring(0, 31)).toEqual(
-          'Karol Miner 336 W. Chugalug Way',
-        );
-        done();
-      });
-    });
-
-    it('can handle funky formatting', (done) => {
-      const filePath = path.join(__dirname, 'files', 'Untitleddocument.docx');
-      fromFileWithPath(filePath, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text).toEqual(
-          "this is a test document that won't be extracted properly. ",
-        );
-        done();
-      });
-    });
-
-    it('can handle a huge docx', (done) => {
-      const filePath = path.join(__dirname, 'files', 'LargeLorem.docx');
-      fromFileWithPath(filePath, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text.substring(0, 100)).toEqual(
-          'Hashtag chambray XOXO PBR&B chia small batch. Before they sold out banh mi raw denim, fap synth hell',
-        );
-        done();
-      });
-    });
-
-    it('can handle arabic', (done) => {
-      const filePath = path.join(__dirname, 'files', 'arabic.docx');
-      fromFileWithPath(filePath, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text.substring(0, 100)).toEqual(
-          ' التعرف الضوئي على الحروف إشعار عدم التمييز (المصدر: مكتب الصحة والخدمات الإنسانية من أجل الحقوق الم',
-        );
-        done();
-      });
-    });
-  });
-
-  describe('for text/* files', () => {
-    it('will extract text from specifically a .txt file', (done) => {
-      const filePath = path.join(__dirname, 'files', 'txt.txt');
-      fromFileWithPath(filePath, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text).to.eql('This is a plain old text file.');
-        done();
-      });
-    });
-
-    it('will extract text from specifically a non utf8 .txt file', (done) => {
-      const filePath = path.join(__dirname, 'files', 'non-utf8.txt');
-      fromFileWithPath(filePath, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text).to.eql('これは非UTF8 テキストファイルです ');
-        done();
-      });
-    });
-
-    it('will error when .txt file encoding cannot be detected', (done) => {
-      const filePath = path.join(__dirname, 'files', 'unknown-encoding.txt');
-      fromFileWithPath(filePath, (error) => {
-        expect(error).not.toBeNull();
-        expect(error.message).toBeInstanceOf(String);
-        expect(error.message).toEqual(
-          'Could not detect encoding for file named [[ unknown-encoding.txt ]]',
-        );
-        done();
-      });
-    });
-
-    it('will extract text specifically from a .css file', (done) => {
-      const filePath = path.join(__dirname, 'files', 'css.css');
-      fromFileWithPath(filePath, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text).toEqual('.foo {color:red}');
-        done();
-      });
-    });
-
-    it('will extract text specifically from a .js file', (done) => {
-      const filePath = path.join(__dirname, 'files', 'js.js');
-      fromFileWithPath(filePath, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text).toEqual('console.log("javascript is cooler than you")');
-        done();
-      });
-    });
-
-    it('will remove extraneous white space from a .txt file', (done) => {
-      const filePath = path.join(__dirname, 'files', 'spacey.txt');
-      fromFileWithPath(filePath, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text).toEqual('this has lots of space');
-        done();
-      });
-    });
-
-    it('will not remove fancy quotes from a .txt file', (done) => {
-      const filePath = path.join(__dirname, 'files', 'fancyquote.txt');
-      fromFileWithPath(filePath, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text).toEqual('this has "fancy" quotes');
-        done();
-      });
-    });
-  });
-
-  describe('for .dxf files', () => {
-    it('will extract text from actual dxf files', (done) => {
-      const filePath = path.join(__dirname, 'files', 'dxf.dxf');
-      fromFileWithPath(filePath, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-
-        expect(text).toEqual(
-          ' PART: FILE: {\fTimes New Roman|b0|i0|c0|p18;(800) 433-1119} {\fTimes New Roman|b0|i0|c0|p18;Barium Springs, NC 28010} {\fTimes New Roman|b0|i0|c0|p18;MultiDrain Systems, Inc.} {\fTimes New Roman|b0|i0|c0|p18;Manufacturers of MultiDrain & EconoDrain } to others for manufacturing or for any other purpose except as specifically authorized in writing by MultiDrain Systems, Inc. Proprietary rights of MultiDrain Systems, Inc. are included in the information disclosed herein. The recipient, by accepting this document, agrees that neither this document nor the information disclosed herein nor any part thereof shall be copied, reproduced or transferred 0 2" 4" 6" 8" 12" 16" GRAPHIC SCALE BAR A1;T A1;T A1;T A1;6.1" 155mm A1;T A1;T A1;4.9" 124mm A1;19.6" 497mm FRAME AND GRATE LENGTH A1;5.5" 140mm %%UCROSS SECTIONAL VIEW SOIL SUBGRADE CONCRETE THICKNESS AND REINFORCEMENT PER STRUCTURAL ENGINEER S SPECIFICATION FOR THE APPLICATION FLOOR SLAB THICKNESS, OR 4" MIN. [100mm], OR SPECIFICATION (WHICHEVER IS GREATER) T = MONOLITHIC CONCRETE POUR (ACCEPTABLE) EXPANSION JOINT BOTH SIDES (PREFERRED) LOCK DOWN BOLT LOCK TOGGLE ANCHOR BOLT SEE ABOVE FOR ACTUAL FRAME & GRATE SECTIONS %%UPLAN %%USECTION 512AF %%UPLAN %%USECTION 513AF 514AF %%UPLAN %%USECTION 515AF %%UPLAN %%USECTION ANCHOR RIB INDEPENDENTLY ANCHORED FRAME ALFA CHANNEL A1;502 GRATE 510AF ANCHOR FRAME 503 GRATE 510AF ANCHOR FRAME 504 GRATE 505 GRATE FRAME AND GRATE ADD 1.2" [31mm] TO OVERALL DEPTH OF CHANNEL LNOTE: GRATE WIDTH FRAME WIDTH AC-2510AF-00 2512AF 2513AF 2514AF 2515AF ALFA CHANNEL SYSTEM DUCTILE IRON FRAME & GRATES PRODUCT DRAWING 2006 MultiDrain Systems, Inc. ',
-        );
-        done();
-      });
-    });
-
-    it('will error when input file is not an actual dxf file', (done) => {
-      const filePath = path.join(__dirname, 'files', 'notadxf.dxf');
-      fromFileWithPath(filePath, (error) => {
-        expect(error).not.toBeNull();
-        expect(error.message).toBeInstanceOf(String);
-        expect(error.message.substring(0, 40)).toEqual(
-          'Error for type: [[ image/vnd.dxf ]], fil',
-        );
-        done();
-      });
-    });
-  });
-
-  describe('for .pptx files', () => {
-    it('will extract text PPTX files', (done) => {
-      const filePath = path.join(__dirname, 'files', 'ppt.pptx');
-      fromFileWithPath(filePath, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text.substring(55, 96)).toEqual(
-          'ullet 1 Bullet 2 Bullet 3 Number 1 Number',
-        );
-        done();
-      });
-    });
-
-    it('will extract text PPTX files with notes', (done) => {
-      const filePath = path.join(__dirname, 'files', 'PrezoWithNotes.pptx');
-      fromFileWithPath(filePath, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text).to.eql('This is a slide These are speaker notes 1 ');
-        done();
-      });
-    });
-
-    it('will extract slides in the right order', (done) => {
-      const filePath = path.join(__dirname, 'files', 'order.pptx');
-      fromFileWithPath(
-        filePath,
-        { preserveLineBreaks: true },
-        (error, text) => {
-          expect(error).toBeNull();
-          expect(text).to.be.an('string');
-          const lines = text.split('\n').filter((line) => /^Slide/.exec(line));
-
-          const linesAnswer = [
-            'Slide 1 Title',
-            'Slide 1 Subtitle',
-            'Slide 2: Title and Content',
-            'Slide 3: Section header',
-            'Slide 4: Two-Content',
-            'Slide 5: Comparison',
-            'Slide 8: Content w/Caption',
-            'Slide 9: picture with caption',
-            'Slide 10: Vertical Text',
-            'Slide 11: Vertical Title and text',
-          ];
-
-          expect(lines).to.eql(linesAnswer);
-
-          done();
-        },
-      );
-    });
-
-    it('will keep preserved characters', (done) => {
-      const filePath = path.join(__dirname, 'files', 'order.pptx');
-      fromFileWithPath(
-        filePath,
-        { preserveLineBreaks: true },
-        (error, text) => {
-          expect(error).toBeNull();
-          expect(text).toBeInstanceOf(String);
-          expect(text.indexOf('…')).toEqual(928);
-          done();
-        },
-      );
-    });
-  });
-
-  describe('for odt files', () => {
-    it('will extract text from ODT files', (done) => {
-      const filePath = path.join(__dirname, 'files', 'spaced.odt');
-      fromFileWithPath(filePath, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text).toEqual('This Is some text');
-        done();
-      });
-    });
-  });
-
-  describe('for image files', () => {
-    it('will extract text from PNG files', (done) => {
-      const filePath = path.join(__dirname, 'files', 'testphoto.png');
-      fromFileWithPath(filePath, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text.substring(0, 100)).toEqual(
-          'performance measure against standards and targets is increasingly used in the management of complex ',
-        );
-        done();
-      });
-    });
-
-    it('will extract text from JPG files', (done) => {
-      const filePath = path.join(__dirname, 'files', 'testphoto.jpg');
-      fromFileWithPath(filePath, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text.substring(0, 100)).toEqual(
-          'performance measure against standards and targets is increasingly used in the management of complex ',
-        );
-        done();
-      });
-    });
-
-    it('will extract text from GIF files', (done) => {
-      const filePath = path.join(__dirname, 'files', 'testphoto.gif');
-      fromFileWithPath(filePath, (error, text) => {
-        expect(error).toBeNull();
-        expect(text).toBeInstanceOf(String);
-        expect(text.substring(0, 100)).toEqual(
-          'performance measure against standards and targets is increasingly used in the management of complex ',
-        );
-        done();
-      });
-    });
-
-    // sudo port install tesseract-chi-sim
-    it('will extract text from language-d files', { timeout: 5000 }, (done) => {
-      const filePath = path.join(__dirname, 'files', 'chi.png');
-
-      fromFileWithPath(
-        filePath,
-        { tesseract: { lang: 'chi_sim' } },
-        (error, text) => {
-          expect(error).toBeNull();
-          expect(text).toBeInstanceOf(String);
-          expect(text.substring(0, 6)).toEqual('卧虎藏龙，卧');
-          done();
-        },
-      );
-    });
-
-    // sudo port install tesseract-eng
-    it('will take tesseract.cmd option', { timeout: 5000 }, (done) => {
-      const filePath = path.join(__dirname, 'files', 'testpng.png');
-      fromFileWithPath(
-        filePath,
-        { tesseract: { cmd: '-l eng -psm 3' } },
-        (error, text) => {
-          expect(error).toBeNull();
-          expect(text).toBeInstanceOf(String);
-          expect(text.substring(0, 100)).toEqual(
-            'The (quick) [brown] {fox} jumps! Over the $43,456.78 <lazy> #90 dog & duck/goose, as 12.5% of E-mail',
-          );
-          done();
-        },
-      );
-    });
-  });
-
-  function test(ext, name, text1, text2) {
-    describe(`for ${ext} files`, () => {
-      it('will extract text', (done) => {
-        const filePath = path.join(__dirname, 'files', name);
-        fromFileWithPath(filePath, (error, text) => {
-          expect(error).toBeNull();
-          expect(text).toBeInstanceOf(String);
-          expect(text.substring(0, 100)).to.eql(text1);
-          done();
-        });
-      });
-
-      it('will extract text and preserve line breaks', (done) => {
-        const filePath = path.join(__dirname, 'files', name);
-        fromFileWithPath(
-          filePath,
-          { preserveLineBreaks: true },
-          (error, text) => {
-            expect(error).toBeNull();
-            expect(text).toBeInstanceOf(String);
-            expect(text.substring(0, 100)).toEqual(text2);
-            done();
-          },
-        );
-      });
-    });
-  }
-
-  test(
-    'markdown',
-    'test.md',
-    ' This is an h1 This is an h2 This text has been bolded and italicized ',
-    '\nThis is an h1\nThis is an h2\nThis text has been bolded and italicized\n',
-  );
-
-  test(
-    'ods',
-    'ods.ods',
-    'This,is,a,ods Really,it,is, I,promise,, ',
-    'This,is,a,ods\nReally,it,is,\nI,promise,,\n',
-  );
-
-  test(
-    'xml',
-    'xml.xml',
-    ' Empire Burlesque Bob Dylan USA Columbia 10.90 1985 Hide your heart Bonnie Tyler UK CBS Records 9.90',
-    '\nEmpire Burlesque\nBob Dylan\nUSA\nColumbia\n10.90\n1985\nHide your heart\nBonnie Tyler\nUK\nCBS Records\n9.90',
-  );
-
-  test(
-    'odt',
-    'odt.odt',
-    'This is an ODT THIS IS A HEADING More ODT',
-    'This is an ODT\nTHIS IS A HEADING\nMore ODT',
-  );
-
-  test(
-    'potx',
-    'potx.potx',
-    'This is a potx template Yep, a potx I had no idea These were even a thing ',
-    'This is a potx template\nYep, a potx\nI had no idea \nThese were even a thing\n',
-  );
-
-  test(
-    'xltx',
-    'xltx.xltx',
-    ',,,,,, Packing Slip ,Your Company Name,,,,"July 24, 2015", , Your Company Slogan,,,,, ,,,,,, ,Addres',
-    ',,,,,, Packing Slip\n,Your Company Name,,,,"July 24, 2015",\n, Your Company Slogan,,,,,\n,,,,,,\n,Addres',
-  );
-
-  test(
-    'ott',
-    'ott.ott',
-    'This is a document template, yay templates! Woo templates get me so excited!',
-    'This is a document template, yay templates!\nWoo templates get me so excited!',
-  );
-
-  test(
-    'ots',
-    'ots.ots',
-    "This,is , template, an,open,office,template isn't,it,awesome?, you,know,it,is ",
-    "This,is , template,\nan,open,office,template\nisn't,it,awesome?,\nyou,know,it,is\n",
-  );
-
-  test(
-    'odg',
-    'odg.odg',
-    "This is a drawing? A drawing, a drawing! This is a drawing, Aren't you mad envious?",
-    "This is a drawing?\nA drawing, a drawing!\nThis is a drawing,\nAren't you mad envious?",
-  );
-
-  test(
-    'otg',
-    'otg.otg',
-    'This is a drawing template A drawing template. Who would really ever need to extract from one of the',
-    'This is a drawing template\nA drawing template.\nWho would really ever need to extract from one of the',
-  );
-
-  test(
-    'odp',
-    'odp.odp',
-    "This is a title This is a slide's text This is a 2nd page And a 2nd page's content",
-    "This is a title\nThis is a slide's text\nThis is a 2nd page\nAnd a 2nd page's content",
-  );
-
-  test(
-    'otp',
-    'otp.otp',
-    'This is a template title Template page text 2nd prezo text',
-    'This is a template title\nTemplate page text\n2nd prezo text',
-  );
-});
