--- test/url.test.js
@@ -1,133 +1,0 @@
-import { describe, it, expect } from 'vitest';
-import { fromUrl } from '../lib/index.js';
-
-const TEST_CASES = [
-  [
-    'doc',
-    'doc.doc',
-    ' Word Specification Sample Working Draft 04, 16 August 2002 Document identifier: wd-spectools-word-s',
-  ],
-
-  ['xls', 'test.xls', 'This,is,a,spreadsheet,yay! '],
-
-  ['xlsx', 'pi.xlsx', 'This is the value of PI:,3.141592 '],
-
-  ['pdf', 'pdf.pdf', 'This is a test. Please ignore.'],
-
-  [
-    'docx',
-    'docx.docx',
-    'This is a test Just so you know: Lorem ipsum dolor sit amet, consecutuer adipiscing elit, sed diam n',
-  ],
-
-  ['text/*', 'txt.txt', 'This is a plain old text file.'],
-
-  [
-    'pptx',
-    'ppt.pptx',
-    'This is some title Text And a sub-title Text in Lists Bullet 1 Bullet 2 Bullet 3 Number 1 Number 2 N',
-  ],
-
-  [
-    'markdown',
-    'test.md',
-    ' This is an h1 This is an h2 This text has been bolded and italicized ',
-  ],
-
-  ['ods', 'ods.ods', 'This,is,a,ods Really,it,is, I,promise,, '],
-
-  [
-    'xml',
-    'xml.xml',
-    ' Empire Burlesque Bob Dylan USA Columbia 10.90 1985 Hide your heart Bonnie Tyler UK CBS Records 9.90',
-  ],
-
-  ['odt', 'odt.odt', 'This is an ODT THIS IS A HEADING More ODT'],
-
-  [
-    'potx',
-    'potx.potx',
-    'This is a potx template Yep, a potx I had no idea These were even a thing ',
-  ],
-
-  [
-    'xltx',
-    'xltx.xltx',
-    ',,,,,, Packing Slip ,Your Company Name,,,,"July 24, 2015", , Your Company Slogan,,,,, ,,,,,, ,Addres',
-  ],
-
-  [
-    'ott',
-    'ott.ott',
-    'This is a document template, yay templates! Woo templates get me so excited!',
-  ],
-
-  [
-    'ots',
-    'ots.ots',
-    "This,is , template, an,open,office,template isn't,it,awesome?, you,know,it,is ",
-  ],
-
-  [
-    'odg',
-    'odg.odg',
-    "This is a drawing? A drawing, a drawing! This is a drawing, Aren't you mad envious?",
-  ],
-
-  [
-    'otg',
-    'otg.otg',
-    'This is a drawing template A drawing template. Who would really ever need to extract from one of the',
-  ],
-
-  [
-    'odp',
-    'odp.odp',
-    "This is a title This is a slide's text This is a 2nd page And a 2nd page's content",
-  ],
-
-  [
-    'otp',
-    'otp.otp',
-    'This is a template title Template page text 2nd prezo text',
-  ],
-];
-
-describe('fromUrl tests', { timeout: 3000 }, () => {
-  it('will properly extract files from sites with extensions that are misleading', (done) => {
-    const url = 'http://apps.leg.wa.gov/billinfo/summary.aspx?bill=1276';
-    fromUrl(url, (error, text) => {
-      expect(error).toBeNull();
-      expect(text).toBeInstanceOf(String);
-      expect(text.substring(0, 100)).toEqual(
-        ' Washington State Legislature Bill Summary 2017-2018 2015-2016 2013-2014 2011-2012 2009-2010 2007-20',
-      );
-      done();
-    });
-  });
-
-  it('take object URL', (done) => {
-    const url =
-      'https://cdn.rawgit.com/dbashford/textract/master/test/files/doc.doc?raw=true';
-    const urlObj = new URL(url);
-    fromUrl(urlObj, (error, text) => {
-      expect(error).toBeNull();
-      expect(text).toBeInstanceOf(String);
-      expect(text.substring(0, 100)).toEqual(
-        ' Word Specification Sample Working Draft 04, 16 August 2002 Document identifier: wd-spectools-word-s',
-      );
-      done();
-    });
-  });
-
-  it.each(TEST_CASES)('will %s files', (_ext, name, text) => {
-    const url = `https://cdn.rawgit.com/dbashford/textract/master/test/files/${
-      name
-    }?raw=true`;
-    fromUrl(url, (error, _text) => {
-      expect(error).toBeNull();
-      expect(text).toBeInstanceOf(String);
-      expect(text.substring(0, 100)).toEqual(text);
-    });
-  });
-});
